Add files via upload
this robot will use 2 IR sensors to follow a black line on a white surface as the main road and 2 IR sensors to count black stripes (or obstacles) on the sides of the road. also it connects to your device using Bluetooth sensor and sends information about the number of lines every second. finally when the path ends it will send a total number of all the lines that it has identified.